requirements.txt
numpy
opencv-python
matplotlib
pyyaml